+++ packages/engine/Source/Scene/Atmosphere.js
@@ -1,4 +1,5 @@
 import Cartesian3 from "../Core/Cartesian3.js";
+import CesiumMath from "../Core/Math.js";
 import DynamicAtmosphereLightingType from "./DynamicAtmosphereLightingType.js";
 
 /**
@@ -123,5 +124,26 @@ function Atmosphere() {
    */
   this.dynamicLighting = DynamicAtmosphereLightingType.NONE;
 }
+
+/**
+ * Returns <code>true</code> if the atmosphere shader requires a color correct step.
+ * @param {Atmosphere} atmosphere The atmosphere instance to check
+ * @returns {boolean} true if the atmosphere shader requires a color correct step
+ */
+Atmosphere.requiresColorCorrect = function (atmosphere) {
+  return !(
+    CesiumMath.equalsEpsilon(atmosphere.hueShift, 0.0, CesiumMath.EPSILON7) &&
+    CesiumMath.equalsEpsilon(
+      atmosphere.saturationShift,
+      0.0,
+      CesiumMath.EPSILON7,
+    ) &&
+    CesiumMath.equalsEpsilon(
+      atmosphere.brightnessShift,
+      0.0,
+      CesiumMath.EPSILON7,
+    )
+  );
+};
 
 export default Atmosphere;