+++ packages/engine/Specs/Scene/ShadowMapSpec.js
@@ -11,6 +11,7 @@ import {
   HeadingPitchRange,
   HeadingPitchRoll,
   HeightmapTerrainData,
+  ImageBasedLighting,
   JulianDate,
   Math as CesiumMath,
   Matrix4,
@@ -281,7 +282,29 @@ describe(
     }
 
     async function loadModel(options) {
-      const model = scene.primitives.add(await Model.fromGltfAsync(options));
+      // A white ambient light with low intensity
+      const defaultIbl = new ImageBasedLighting({
+        sphericalHarmonicCoefficients: [
+          new Cartesian3(0.35449, 0.35449, 0.35449),
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+          Cartesian3.ZERO,
+        ],
+      });
+      const model = scene.primitives.add(
+        await Model.fromGltfAsync({
+          environmentMapOptions: {
+            enabled: false,
+          },
+          imageBasedLighting: defaultIbl,
+          ...options,
+        }),
+      );
       await pollToPromise(
         function () {
           // Render scene to progressively load the model